Save word, song_url pairs in redis
The scraper checks if a url for a given word is saved in redis. If this
is the case, this url is used to create the Song object instead of
querying the SoundCloud site.

This could expanded to the embed_url as well. However, I then had to
serialize an object instead of saving key, value pairs of word and url.